// src/frontend/src/api/schema.d.ts
export interface components {
    schemas: {
// ...
        SuspectDossierOut: {
            /** Suspect Id */
            suspect_id: number;
            /** Name */
            name: string | null;
            /** Aliases */
            aliases: string[];
            /** Phone Numbers */
            phone_numbers: string[];
            /** Vehicle Numbers */
            vehicle_numbers: string[];
            /** Physical Description */
            physical_description: string | null;
            /** Associated Fir Ids */
            associated_fir_ids: number[];
            cluster: components["schemas"]["OffenderClusterOut"] | null;
        };
        /** ValidationError */
        ValidationError: {
            /** Location */
            loc: (string | number)[];
            /** Message */
            msg: string;
            /** Error Type */
            type: string;
        };
    };
    responses: never;
    parameters: never;
    requestBodies: never;
    headers: never;
    pathItems: never;
}